Renomenando path do script, criando mas duas extrações da base, total de beneficiados registrados e anonimos

+++ Auxilio_Emergencial/auxilio_emerg_spark.py
@@ -1,6 +1,7 @@
+from genericpath import exists
 import os
 from pyspark.sql import SparkSession
-from pyspark.sql.functions import regexp_replace, col
+from pyspark.sql.functions import regexp_replace, col, count
 
 def tratamento_csv(path_file :str, destino:str):
     spark = SparkSession.builder.getOrCreate()
@@ -50,7 +51,22 @@ def tratamento_csv(path_file :str, destino:str):
     csv_municipio = path_destino+'\\'+'municipios.csv'
     municipios.toPandas().to_csv(csv_municipio)
 
-    if os.path.exists(file_csv) and os.path.exists(csv_municipio):
+    total_beneficiados = df2.filter("COD_MUNICIPIO IS NOT NULL AND CPF_BENEF IS NOT NULL")\
+                            .groupBy(['ANO_MES','COD_MUNICIPIO'])\
+                            .agg(count('COD_MUNICIPIO').alias('TOTAL_BENEF'))
+
+    csv_beneficiados = path_destino+'\\'+'beneficiados_registrados.csv'
+    total_beneficiados.toPandas().to_csv(csv_beneficiados)
+
+    total_anonimos = df2.filter("COD_MUNICIPIO IS NOT NULL AND CPF_BENEF IS NULL")\
+                            .groupBy(['ANO_MES','COD_MUNICIPIO'])\
+                            .agg(count('COD_MUNICIPIO').alias('TOTAL_ANONIMOS'))
+
+    csv_anonimos = path_destino+'\\'+'beneficiados_anonimos.csv'
+    total_anonimos.toPandas().to_csv(csv_anonimos)
+
+    if os.path.exists(file_csv) and os.path.exists(csv_municipio)\
+        and os.path.exists(csv_beneficiados) and os.path.exists(csv_anonimos):
         os.remove(path_file)
         return True
     return False
